# lib/api-zod/src/generated/types/instrument.ts
/**
 * Generated by orval v8.23.0 🍺
 * Do not edit manually.
 * Api
 * Broker-independent trading strategy workspace API
 * OpenAPI spec version: 0.1.0
 */
import type { InstrumentInstrumentType } from './instrumentInstrumentType';

export interface Instrument {
  id: number;
  assetClass: string;
  instrumentType: InstrumentInstrumentType;
  /** @nullable */
  venue: string | null;
  symbol: string;
  /** @nullable */
  displayName: string | null;
  /** @nullable */
  baseCurrency: string | null;
  /** @nullable */
  quoteCurrency: string | null;
  /** @nullable */
  exchangeTimezone: string | null;
  /** @nullable */
  tickSize: number | null;
  /** @nullable */
  contractMultiplier: number | null;
  /** @nullable */
  expiry: string | null;
  isActive: boolean;
  /** @nullable */
  description: string | null;
  createdAt: Date;
  updatedAt: Date;
}