# lib/api-zod/src/generated/types/instrumentUpdate.ts
/**
 * Generated by orval v8.23.0 🍺
 * Do not edit manually.
 * Api
 * Broker-independent trading strategy workspace API
 * OpenAPI spec version: 0.1.0
 */
import type { InstrumentUpdateInstrumentType } from './instrumentUpdateInstrumentType';

export interface InstrumentUpdate {
  /** @minLength 1 */
  assetClass?: string;
  instrumentType?: InstrumentUpdateInstrumentType;
  /** @nullable */
  venue?: string | null;
  /** @minLength 1 */
  symbol?: string;
  /** @nullable */
  displayName?: string | null;
  /** @nullable */
  baseCurrency?: string | null;
  /** @nullable */
  quoteCurrency?: string | null;
  /** @nullable */
  exchangeTimezone?: string | null;
  /**
     * @exclusiveMinimum 0
     * @nullable
     */
  tickSize?: number | null;
  /**
     * @exclusiveMinimum 0
     * @nullable
     */
  contractMultiplier?: number | null;
  /** @nullable */
  expiry?: string | null;
  isActive?: boolean;
  /** @nullable */
  description?: string | null;
}