# lib/api-zod/src/generated/types/marketDataSource.ts
/**
 * Generated by orval v8.23.0 🍺
 * Do not edit manually.
 * Api
 * Broker-independent trading strategy workspace API
 * OpenAPI spec version: 0.1.0
 */
import type { MarketDataSourceCapabilitiesItem } from './marketDataSourceCapabilitiesItem';
import type { MarketDataSourceConfigurationStatus } from './marketDataSourceConfigurationStatus';
import type { MarketDataSourceSourceType } from './marketDataSourceSourceType';

export interface MarketDataSource {
  id: number;
  name: string;
  /** @nullable */
  providerKey: string | null;
  sourceType: MarketDataSourceSourceType;
  /** @nullable */
  description: string | null;
  capabilities: MarketDataSourceCapabilitiesItem[];
  configurationStatus: MarketDataSourceConfigurationStatus;
  isEnabled: boolean;
  createdAt: Date;
  updatedAt: Date;
}