# lib/api-zod/src/generated/types/timeframe.ts
/**
 * Generated by orval v8.23.0 🍺
 * Do not edit manually.
 * Api
 * Broker-independent trading strategy workspace API
 * OpenAPI spec version: 0.1.0
 */

export interface Timeframe {
  id: number;
  code: string;
  label: string;
  durationSeconds: number;
  /** @nullable */
  description: string | null;
  isActive: boolean;
  createdAt: Date;
  updatedAt: Date;
}