# lib/api-zod/src/generated/types/timeframeInput.ts
/**
 * Generated by orval v8.23.0 🍺
 * Do not edit manually.
 * Api
 * Broker-independent trading strategy workspace API
 * OpenAPI spec version: 0.1.0
 */

export interface TimeframeInput {
  /** @minLength 1 */
  code: string;
  /** @minLength 1 */
  label: string;
  /** @minimum 1 */
  durationSeconds: number;
  /** @nullable */
  description?: string | null;
  isActive?: boolean;
}